requirements.txt
groq
langchain==0.1.16
langchain-core
langchain-groq
gunicorn
Flask
line-bot-sdk
python-dotenv